requirements.txt
asgiref==3.4.1
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
django-heroku
psycopg2-binary==2.9.1
python-decouple==3.5
pytz==2021.3
sqlparse==0.4.2
Shapely
euclid3
solidpython
boto3
numpy
gunicorn
whitenoise